Rename: my 폴더 이름 mypage로 변경
