+++ src/components/mypage/events/StatusItem.tsx
