+++ src/components/mypage/events/input/EventDateInput.tsx
@@ -2,7 +2,9 @@ import { UseFormReturn } from 'react-hook-form';
 import { format } from 'date-fns';
 import { ko } from 'date-fns/locale';
 import { CalendarIcon } from 'lucide-react';
+import { Select } from '@radix-ui/react-select';
 
+import { Button } from '@components/ui/button';
 import {
   FormControl,
   FormField,
@@ -14,10 +16,8 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from '@components/ui/popover';
-import { Button } from '@components/ui/button';
 import { Calendar } from '@components/ui/calendar';
 import {
-  Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
@@ -28,38 +28,30 @@ import EventInputTitle from '../EventInputTitle';
 import { EventFormValues } from '@/types/form';
 import { cn } from '@/lib/utils';
 import RightArrow from '@assets/images/icons/RightArrow.svg';
-import formatTimeTo12HourClock from '@/utils/my/formatTimeTo12HourClock';
+import formatTimeTo12HourClock from '@utils/mypage/formatTimeTo12HourClock';
 
-interface EventRegistrationDateInputProps {
+interface EventDateInputProps {
   form: UseFormReturn<EventFormValues>;
   startDate: Date | null;
-  registrationStartDate: Date | null;
-  registrationStartTime: string;
+  startTime: string;
+  registrationEndDate: Date | null;
 }
 
-function EventRegistrationDateInput({
+function EventDateInput({
   form,
   startDate,
-  registrationStartDate,
-  registrationStartTime,
-}: EventRegistrationDateInputProps) {
+  startTime,
+  registrationEndDate,
+}: EventDateInputProps) {
   return (
     <div className='flex flex-col space-y-2'>
-      <EventInputTitle
-        title={
-          <>
-            모집 기간{' '}
-            <span className='text-xs font-normal text-gray-500'>
-              (이벤트 일시를 먼저 선택해 주세요.)
-            </span>
-          </>
-        }
-      />
+      <EventInputTitle title='이벤트 일시' />
+
       <div className='flex flex-wrap'>
         <div className='flex-grow basis-[22%]'>
           <FormField
             control={form.control}
-            name='registrationStartDate'
+            name='startDate'
             render={({ field }) => (
               <FormItem className='flex flex-grow flex-col'>
                 <Popover>
@@ -87,10 +79,12 @@ function EventRegistrationDateInput({
                       selected={field.value ? new Date(field.value) : undefined}
                       onSelect={field.onChange}
                       disabled={date => {
-                        const startDateValue = startDate
-                          ? new Date(startDate)
-                          : new Date();
-                        return date < new Date() || date >= startDateValue;
+                        const registrationEndDateValue = registrationEndDate
+                          ? new Date(registrationEndDate)
+                          : null;
+                        return registrationEndDateValue
+                          ? date <= registrationEndDateValue
+                          : new Date() > date;
                       }}
                       initialFocus
                     />
@@ -106,7 +100,7 @@ function EventRegistrationDateInput({
           <FormField
             control={form.control}
             rules={{ required: '시작 시간은 필수입니다.' }}
-            name='registrationStartTime'
+            name='startTime'
             render={({ field }) => (
               <FormItem className='flex flex-grow flex-col'>
                 <Select
@@ -154,7 +148,7 @@ function EventRegistrationDateInput({
         <div className='flex-grow basis-[22%]'>
           <FormField
             control={form.control}
-            name='registrationEndDate'
+            name='endDate'
             render={({ field }) => (
               <FormItem className='flex flex-grow flex-col'>
                 <Popover>
@@ -181,18 +175,11 @@ function EventRegistrationDateInput({
                       mode='single'
                       selected={field.value ? new Date(field.value) : undefined}
                       onSelect={field.onChange}
-                      disabled={date => {
-                        const registrationStartDateValue = registrationStartDate
-                          ? new Date(registrationStartDate)
-                          : new Date();
-                        const eventStartDateValue = startDate
-                          ? new Date(startDate)
-                          : new Date();
-                        return (
-                          date < registrationStartDateValue ||
-                          date >= eventStartDateValue
-                        ); // 모집 시작 날짜 이후이면서 이벤트 시작 날짜 이전 날짜만 선택 가능
-                      }}
+                      disabled={date =>
+                        startDate
+                          ? date < new Date(startDate)
+                          : date < new Date()
+                      }
                       initialFocus
                     />
                   </PopoverContent>
@@ -207,13 +194,13 @@ function EventRegistrationDateInput({
           <FormField
             control={form.control}
             rules={{ required: '종료 시간은 필수입니다.' }}
-            name='registrationEndTime'
+            name='endTime'
             render={({ field }) => (
               <FormItem className='flex flex-grow flex-col'>
                 <Select
                   onValueChange={field.onChange}
                   defaultValue={field.value}
-                  disabled={!registrationStartDate || !registrationStartTime}
+                  disabled={!startDate || !startTime}
                 >
                   <FormControl>
                     <SelectTrigger>
@@ -253,4 +240,4 @@ function EventRegistrationDateInput({
   );
 }
 
-export default EventRegistrationDateInput;
+export default EventDateInput;
