+++ src/components/mypage/events/input/EventThumbnailInput.tsx
