+++ src/components/mypage/events/input/EventTicketInput.tsx
