+++ src/pages/mypage/MyEvents.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useQuery } from 'react-query';
 
-import Status from '@components/my/events/Status';
-import EventTable from '@components/my/events/EventTable';
+import Status from '@components/mypage/events/Status';
+import EventTable from '@components/mypage/events/EventTable';
 import { Button } from '@components/ui/button';
 import SpinnerBox from '@components/shared/SpinnerBox';
 import ErrorBox from '@components/shared/ErrorBox';
@@ -44,7 +44,7 @@ function MyEvents() {
       {/* <button onClick={addDummyEvents}>더미데이터 추가</button> */}
       <div className='mb-4 flex justify-end'>
         <Button>
-          <Link to='/my/events/new'>등록하기</Link>
+          <Link to='/mypage/events/new'>등록하기</Link>
         </Button>
       </div>
       <div>
