+++ src/utils/mypage/calculateEventStatus.ts
